requirements.txt
annotated-types==0.6.0
greenlet==3.0.3
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
typing_extensions==4.11.0
annotated-types==0.6.0
docstring_parser==0.16
greenlet==3.0.3
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
typing_extensions==4.11.0
annotated-types==0.6.0
docstring_parser==0.16
greenlet==3.0.3
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
PyYAML==6.0.1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
typing_extensions==4.11.0
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
docstring_parser==0.16
greenlet==3.0.3
idna==3.7
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
typing_extensions==4.11.0
urllib3==2.2.1
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
docstring_parser==0.16
greenlet==3.0.3
idna==3.7
numpy==1.26.4
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
typing_extensions==4.11.0
urllib3==2.2.1
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
docstring_parser==0.16
greenlet==3.0.3
idna==3.7
numpy==1.26.4
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
tiktoken==0.6.0
typing_extensions==4.11.0
urllib3==2.2.1
